Extract VideoUrlValidator
